// Back-end/src/domain/library/models/book-languages.model.ts
export enum BookLanguagesModel {
  English = "English",
  Chinese = "Chinese",
  Hindi = "Hindi",
  Spanish = "Spanish",
  French = "French",
  Arabic = "Arabic",
  Bengali = "Bengali",
  Portuguese = "Portuguese",
  Russian = "Russian",
  Urdu = "Urdu",
  Indonesian = "Indonesian",
  German = "German",
  Japanese = "Japanese",
  Swahili = "Swahili",
  Marathi = "Marathi",
  Telugu = "Telugu",
  Turkish = "Turkish",
  Korean = "Korean",
  Tamil = "Tamil",
  Vietnamese = "Vietnamese",
}